refactor: split the styled-components interfaces in bespoke files
