+++ src/typings/styledComponents/buttons.d.ts
@@ -1,0 +1,7 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface Buttons {
+    size: ButtonsSize;
+  }
+}
