+++ src/typings/styledComponents/buttonsSize.d.ts
@@ -1,0 +1,8 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface ButtonsSize {
+    md: PurpleLg;
+    lg: PurpleLg;
+  }
+}
