+++ src/typings/styledComponents/disabled.d.ts
@@ -1,0 +1,7 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface Disabled {
+    opacity: number;
+  }
+}
