+++ src/typings/styledComponents/fonts.d.ts
@@ -1,0 +1,8 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface Fonts {
+    fontFamily: string;
+    fontFamilyCode: string;
+  }
+}
