+++ src/typings/styledComponents/icon.d.ts
@@ -1,0 +1,7 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface Icon {
+    size: IconTextSize;
+  }
+}
