+++ src/typings/styledComponents/iconTextSize.d.ts
@@ -1,0 +1,8 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface IconTextSize {
+    sm: null | string;
+    md: null | string;
+  }
+}
