+++ src/typings/styledComponents/loader.d.ts
@@ -1,0 +1,7 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface Loader {
+    size: LoaderSize;
+  }
+}
