+++ src/typings/styledComponents/loaderSize.d.ts
@@ -1,0 +1,10 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface LoaderSize {
+    xs: string;
+    sm: string;
+    md: string;
+    lg: string;
+  }
+}
