+++ src/typings/styledComponents/overlay.d.ts
@@ -1,0 +1,8 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface Overlay {
+    opacity: number;
+    color: string;
+  }
+}
