+++ src/typings/styledComponents/purpleLg.d.ts
@@ -1,0 +1,8 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface PurpleLg {
+    height: string;
+    padding: string;
+  }
+}
