+++ src/typings/styledComponents/shadow.d.ts
@@ -1,0 +1,9 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface Shadow {
+    blur: string;
+    opacity: number;
+    color: string;
+  }
+}
