+++ src/typings/styledComponents/smClass.d.ts
@@ -1,0 +1,8 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface SmClass {
+    fontSize: string;
+    lineHeight: string;
+  }
+}
