+++ src/typings/styledComponents/text.d.ts
@@ -1,0 +1,7 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface Text {
+    size: TextSize;
+  }
+}
